requirements.txt
streamlit

# Langchain core
langchain
langchain-core

#Google Gemini Integration
langchain-google-genai
google-generativeai

# Environment variable Management
python-dotenv

#Document Loaders
PyPDF2
